# hocode-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server Api Hocode.",
        "title": "Hocode API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "hocodevn.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/updateusercourse": {
            "post": {
                "description": "UpdateUserCourse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserCourse"
                ],
                "summary": "UpdateUserCourse",
                "parameters": [
                    {
                        "description": "UpdateUserCourse",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BodyUC"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserCourseOut"
                        }
                    }
                }
            }
        },
...
        "/books": {
            "get": {
                "description": "get books \u003ca href=\"/api/v1/books?page=1\u0026limit=5\"\u003e/api/v1/books?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "List Books",
                "operationId": "book_list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Book"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Create Book",
                "operationId": "book_create",
                "parameters": [
                    {
                        "description": "Create Book",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            }
        },
        "/books/:id": {
            "get": {
                "description": "get books \u003ca href=\"/api/v1/books?page=1\u0026limit=5\"\u003e/api/v1/books?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "One Books",
                "operationId": "book_one",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "UpdateBooks Book",
                "operationId": "book_update",
                "parameters": [
                    {
                        "description": "Update Book",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Delete Book",
                "operationId": "book_delete",
                "parameters": [
                    {
                        "description": "Delete Book",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            }
        },
        "/certs": {
            "get": {
                "description": "get certs \u003ca href=\"/api/v1/certs?page=1\u0026limit=5\"\u003e/api/v1/certs?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certs"
                ],
                "summary": "List Certs",
                "operationId": "cert_list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Cert"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Cert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certs"
                ],
                "summary": "Create Cert",
                "operationId": "cert_create",
                "parameters": [
                    {
                        "description": "Create Cert",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                }
            }
        },
        "/certs/:id": {
            "get": {
                "description": "get certs \u003ca href=\"/api/v1/certs?page=1\u0026limit=5\"\u003e/api/v1/certs?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certs"
                ],
                "summary": "One Certs",
                "operationId": "cert_one",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Cert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certs"
                ],
                "summary": "UpdateCerts Cert",
                "operationId": "cert_update",
                "parameters": [
                    {
                        "description": "Update Cert",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Cert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certs"
                ],
                "summary": "Delete Cert",
                "operationId": "cert_delete",
                "parameters": [
                    {
                        "description": "Delete Cert",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cert"
                        }
                    }
                }
            }
        },
...
        "/configs": {
            "get": {
                "description": "get configs \u003ca href=\"/api/v1/configs?page=1\u0026limit=5\"\u003e/api/v1/configs?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configs"
                ],
                "summary": "List Configs",
                "operationId": "config_list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Config"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configs"
                ],
                "summary": "Create Config",
                "operationId": "config_create",
                "parameters": [
                    {
                        "description": "Create Config",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                }
            }
        },
        "/configs/:id": {
            "get": {
                "description": "get configs \u003ca href=\"/api/v1/configs?page=1\u0026limit=5\"\u003e/api/v1/configs?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configs"
                ],
                "summary": "One Configs",
                "operationId": "config_one",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configs"
                ],
                "summary": "UpdateConfigs Config",
                "operationId": "config_update",
                "parameters": [
                    {
                        "description": "Update Config",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configs"
                ],
                "summary": "Delete Config",
                "operationId": "config_delete",
                "parameters": [
                    {
                        "description": "Delete Config",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                }
            }
        },
        "/courses": {
            "get": {
                "description": "get courses \u003ca href=\"/api/v1/courses?page=1\u0026limit=5\"\u003e/api/v1/courses?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "List Courses",
                "operationId": "coursesx_list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Course"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Create Course",
                "operationId": "coursesx_create",
                "parameters": [
                    {
                        "description": "Create Course",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            }
        },
...
        "/courses/{id}": {
            "get": {
                "description": "get courses by ID \u003ca href=\"/courses/5d86e07bfe6e2b157bd3b259\"\u003e/courses/5d86e07bfe6e2b157bd3b259\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Get Course By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.BodyUC": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string"
                },
                "minitask_id": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
        "model.Book": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "del": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.UserCourseOut": {
            "type": "object",
            "properties": {
                "codepoint": {
                    "type": "integer"
                },
                "next_minitask": {
                    "type": "object",
                    "$ref": "#/definitions/model.MiniTask"
                },
                "user_course": {
                    "type": "object",
                    "$ref": "#/definitions/model.UserCourse"
                },
                "user_minitask": {
                    "type": "object",
                    "$ref": "#/definitions/model.UserMiniTask"
                }
            }
        },
...
    }
}